Fix some SBCL optimization warnings
The only warnings that are not fixed are related to SEQUENCE still being a
sequence. This leaves us with three options: either SEQUENCE is a list, or a
vector, or an extended sequence. We will need to generate three different
branches for all three options to get the best possibility for optimization.